Add files via upload
This document provides a step-by-step guide on installing Android-x86, an optimized version of Android for x86 processor-based devices. It explains the filesystem that Android-x86 utilizes. Additionally, it defines virtualization, explaining its purpose, benefits and how it enables the execution of multiple operating systems on a single device. By leveraging virtualization, one can test and run Android-x86 seamlessly like VMware. This document serves as a source for understanding both system installation and the underlying technical concepts.